Refactor discard wireless
Move the discard wireless connections code from the control panel
section where it is used to the network model which it uses heavily.

Simplifies the code.  No functional change.
